requirements.txt
mccode-antlr >= 0.18.7
ncrystal
gitpython
pytest
